Consolidating Functions in DataBase
