+++ ArkansasAssetBuilders/src/main/java/model/Demographic.java
@@ -1,244 +1,112 @@
 package model;
-
 import javafx.beans.property.*;
-
-/**
- * Demographic class
- * <br/>
- * <br/>
- * Demographic information from inputted tax information.
- * Demographic information is stored in the local database.
- */
 public class Demographic {
-    private StringProperty Address;
-    private IntegerProperty Zip;
-    private StringProperty County;
-    private StringProperty State;
-    private Client client;
-    private ReturnData returnData;
-    private TaxYear taxYear;
+	private StringProperty client_id;
+	private DoubleProperty tax_year;
+	private StringProperty city;
+	private StringProperty state;
+	private StringProperty zip;
+	private DoubleProperty primary_secondary_60_plus;
+	private StringProperty residency;
+	public Demographic(){
+		this.client_id = new SimpleStringProperty();
+		this.tax_year = new SimpleDoubleProperty();
+		this.city = new SimpleStringProperty();
+		this.state = new SimpleStringProperty();
+		this.zip = new SimpleStringProperty();
+		this.state = new SimpleStringProperty();
+		this.primary_secondary_60_plus = new SimpleDoubleProperty();
+		this.residency = new SimpleStringProperty();
+	}
+	public String getClient(){ return client_id.get();}
 
-    /**
-     * Constructor for the Demographic class.
-     */
-    public Demographic(){
-        this.Address = new SimpleStringProperty();
-        this.Zip = new SimpleIntegerProperty();
-        this.County = new SimpleStringProperty();
-        this.State = new SimpleStringProperty();
-        this.client = new Client();
-        this.returnData = new ReturnData();
-        this.taxYear = new TaxYear();
-    }
+	public void setClient(String client_id){this.client_idset(client_id);}
 
-    /**
-     * Method for getting the ID of the Demographic object.
-     * @return int, ID of the Demographic.
-     */
-    public String getClient_ID(){
-        return client.getID();
-    }
+	public StringProperty ClientProperty(){return c;}
 
-    /**
-     * Set the ID of the Demographic object.
-     * @param Client_ID int, ID that the Demographic object will be set to.
-     */
-    public void setClient_ID(String Client_ID){
-        client.setID(Client_ID);
-    }
 
-    /**
-     * Method for retrieving Demographic ID from the database in an
-     * IntegerProperty format.<br/>
-     * @return IntegerProperty, the ID of the Demographic.
-     */
-    public StringProperty Client_IDProperty(){
-        return client.Client_IDProperty();
-    }
+	public String getClientId(){ return client_id.get();}
 
-    /**
-     * Method for getting the tax year ID of the Demographic object.
-     * @return int, tax year ID of the Demographic.
-     */
-    public int getTaxYear(){
-        return taxYear.getTaxYear();
-    }
+	public void setClientId(String client_id){this.client_idset(client_id);}
 
-    /**
-     * Method for setting the tax year ID of the Demographic object.
-     * @param TaxYear, int, ID that the tax year ID will be set to.
-     */
-    public void setTaxYear(int TaxYear){
-        taxYear.setTaxYear(TaxYear);
-    }
+	public StringProperty ClientIdProperty(){return c;}
 
-    /**
-     * Method for retrieving Demographic tax year ID from the database
-     * in an IntegerProperty format.<br/>
-     * @return IntegerProperty, the tax year ID of the Demographic.
-     */
-    public IntegerProperty taxYearProperty(){
-        return taxYear.taxYearProperty();
-    }
 
-    /**
-     * Method for getting the address of the Demographic object.
-     * @return String, the Address of the Demographic.
-     */
-    public String getAddress(){
-        return Address.get();
-    }
+	public double getTax(){ return tax_year.get();}
 
-    /**
-     * Method for setting the address of the Demographic.
-     * @param Address String, the new address that the Address will be set to.
-     */
-    public void setAddress(String Address){
-        this.Address.set(Address);
-    }
+	public void setTax(double tax_year){this.tax_yearset(tax_year);}
 
-    /**
-     * Method for retrieving Demographic address from the database
-     * in a StringProperty format.<br/>
-     * @return StringProperty, the Address of the Demographic.
-     */
-    public StringProperty addressProperty(){
-        return Address;
-    }
+	public DoubleProperty TaxProperty(){return c;}
 
-    /**
-     * Method for getting the zip code of the Demographic.
-     * @return int, zip code of the Demographic.
-     */
-    public int getZip(){
-        return Zip.get();
-    }
 
-    /**
-     * Method for setting the zip code of the Demographic object.
-     * @param Zip int, new zip code that will replace the old zip code.
-     */
-    public void setZip(int Zip){
-        this.Zip.set(Zip);
-    }
+	public double getTaxYear(){ return tax_year.get();}
 
-    /**
-     * Method for retrieving Demographic zip code from the database
-     * in an IntegerProperty format.<br/>
-     * @return IntegerProperty, the zip code of the Demographic.
-     */
-    public IntegerProperty zipProperty(){
-        return Zip;
-    }
+	public void setTaxYear(double tax_year){this.tax_yearset(tax_year);}
 
-    /**
-     * Method for getting the county of the Demographic object.
-     * @return String, county of the Demographic object.
-     */
-    public String getCounty(){
-        return County.get();
-    }
+	public DoubleProperty TaxYearProperty(){return c;}
 
-    /**
-     * Method for setting the county of the Demographic object.
-     * @param County String, the new county that will replace the old county.
-     */
-    public void setCounty(String County){
-        this.County.set(County);
-    }
 
-    /**
-     * Method for retrieving Demographic county from the database
-     * in a StringProperty format.<br/>
-     * @return StringProperty, the county of the Demographic.
-     */
-    public StringProperty countyProperty(){
-        return County;
-    }
+	public String getCity(){ return city.get();}
 
-    /**
-     * Method for getting the State of the Demographic object.
-     * @return String, the state of the Demographic object.
-     */
-    public String getState(){
-        return State.get();
-    }
+	public void setCity(String city){this.cityset(city);}
 
-    /**
-     * Method for setting the state of a Demographic object.
-     * @param State String, the new state that will replace the old state.
-     */
-    public void setState(String State){
-        this.State.set(State);
-    }
+	public StringProperty CityProperty(){return c;}
 
-    /**
-     * Method for retrieving Demographic State from the database
-     * in a StringProperty format.<br/>
-     * @return StringProperty, the State of the Demographic.
-     */
-    public StringProperty stateProperty(){
-        return State;
-    }
 
-    public String getFirstName(){return client.getFirstName();}
+	public String getState(){ return state.get();}
 
-    public void setFirstName(String FirstName){client.setFirstName(FirstName);}
+	public void setState(String state){this.stateset(state);}
 
-    public StringProperty firstNameProperty(){return client.firstNameProperty();}
+	public StringProperty StateProperty(){return c;}
 
-    public String getLastName(){return client.getLastName();}
 
-    public void setLastName(String LastName){client.setLastName(LastName);}
+	public String getZip(){ return zip.get();}
 
-    public StringProperty lastNameProperty(){return client.lastNameProperty();}
+	public void setZip(String zip){this.zipset(zip);}
 
-    public String getDoB(){return client.getDoB();}
+	public StringProperty ZipProperty(){return c;}
 
-    public void setDoB(String DoB){client.setDoB(DoB);}
 
-    public StringProperty doBProperty(){return client.doBProperty();}
+	public String getState(){ return state.get();}
 
-    public int getLast4SS(){return client.getLast4SS();}
+	public void setState(String state){this.stateset(state);}
 
-    public void setLast4SS(int Last4SS){client.setLast4SS(Last4SS);}
+	public StringProperty StateProperty(){return c;}
 
-    public IntegerProperty last4SSProperty(){return client.last4SSProperty();}
 
-    public int getFederalReturn(){return returnData.getFederalReturn();}
+	public double getPrimary(){ return primary_secondary_60_plus.get();}
 
-    public void setFederalReturn(int FederalReturn){returnData.setFederalReturn(FederalReturn);}
+	public void setPrimary(double primary_secondary_60_plus){this.primary_secondary_60_plusset(primary_secondary_60_plus);}
 
-    public IntegerProperty federalReturnProperty(){return returnData.federalReturnProperty();}
+	public DoubleProperty PrimaryProperty(){return c;}
 
-    public int getTotalRefund(){return returnData.getTotalRefund();}
 
-    public void setTotalRefund(int TotalRefund){returnData.setTotalRefund(TotalRefund);}
+	public double getPrimarySecondary(){ return primary_secondary_60_plus.get();}
 
-    public IntegerProperty totalRefundProperty(){return returnData.totalRefundProperty();}
+	public void setPrimarySecondary(double primary_secondary_60_plus){this.primary_secondary_60_plusset(primary_secondary_60_plus);}
 
-    public int getEITC(){return returnData.getEITC();}
+	public DoubleProperty PrimarySecondaryProperty(){return c;}
 
-    public void setEITC(int EITC){returnData.setEITC(EITC);}
 
-    public IntegerProperty EITCProperty(){return returnData.EITCProperty();}
+	public double getPrimarySecondary60(){ return primary_secondary_60_plus.get();}
 
-    public int getCTC(){return returnData.getCTC();}
+	public void setPrimarySecondary60(double primary_secondary_60_plus){this.primary_secondary_60_plusset(primary_secondary_60_plus);}
 
-    public void setCTC(int CTC){returnData.setCTC(CTC);}
+	public DoubleProperty PrimarySecondary60Property(){return c;}
 
-    public IntegerProperty CTCProperty(){return returnData.CTCProperty();}
 
-    public int getDependents(){return returnData.getDependents();}
+	public double getPrimarySecondary60Plus(){ return primary_secondary_60_plus.get();}
 
-    public void setDependents(int Dependents){returnData.setDependents(Dependents);}
+	public void setPrimarySecondary60Plus(double primary_secondary_60_plus){this.primary_secondary_60_plusset(primary_secondary_60_plus);}
 
-    public IntegerProperty dependentsProperty(){return returnData.dependentsProperty();}
+	public DoubleProperty PrimarySecondary60PlusProperty(){return c;}
 
-    public int getSurveyScore(){return returnData.getSurveyScore();}
 
-    public void setSurveyScore(int SurveyScore){returnData.setSurveyScore(SurveyScore);}
+	public String getResidency(){ return residency.get();}
 
-    public IntegerProperty surveyScoreProperty(){return returnData.surveyScoreProperty();}
+	public void setResidency(String residency){this.residencyset(residency);}
+
+	public StringProperty ResidencyProperty(){return c;}
+
 
 }
