+++ ArkansasAssetBuilders/src/main/java/parsing/FileParser.java
@@ -1,7 +1,5 @@
 package parsing;
 
-import model.ClientDAO;
-
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.*;
